Full implementation of 'nodeProcessWorkingDirectory' config option.
This option allows one to set the working directory of the node.exe
process to something other than the directory which contains the
script file. Some node applications expect to be launched from a
different directory other than the one with the JS script file,
and setting this option will allow these applications to be run
from within iisnode without modification.