Map HTTPS server variable to X-Forwarded-Proto header
In express.js, "trust proxy" mode allows is to properly handle the
scenario of being behind a proxy server, while still properly exposing
the client IP/protocol from the request object.

The enableXFF configuration option for iisnode propagates the
X-Forwarded-For header, but not X-Forwarded-Proto. This commit adds the
X-Forwarded-Proto header as well, in scenarios where enableXFF is true.

For details, see:
https://github.com/tjanczuk/iisnode/issues/265