X-Forwarded-For shouldn't include port
The WSAAddressToString function used to convert the client IP to a
string was also including the client port in the result. According to
the "spec" for X-Forwarded-For, it is only supposed to include the
numeric IP address.

I've switched the code to use the GetNameInfo function instead, with the
NI_NUMERICHOST parameter. This properly returns just a formatted IP
address, and works for both IPv4 and IPv6 addresses.