Fix unclickable triangles in node-inspector
The version of node-inspector being used is very old, and contains some
calls which no longer work in Chrome. node-inspector itself fixed this
bug in December of 2014, and it's frustrating enough for me that I
ported the changes into iisnode's version of the debugger.

Original fix: https://github.com/node-inspector/node-inspector/pull/479